requirements.txt
requests==2.21.0
speedtest-cli==2.1.1
urllib3==1.24.2
